DA with propeller

The function calculates differential cell abundance the transformed proportions and performs a t-test based on a given design matrix

Usage

propellerCalc(
  seu_obj1,
  condition1,
  condition2,
  cluster_col,
  meta_col,
  lookup,
  sample_col,
  formula,
  min_cells = 30
)

Arguments

seu_obj1

A seurat object

condition1

A character representing the first condition

condition2

A character representing the second condition

cluster_col

A character representing the name of the column which contains cluster information

meta_col

A character representing the name of the column which contains meta information, should be in Seurat object and in lookup table

lookup

A dataframe that contains sample information

sample_col

A character representing the name of the column in seu_obj1 which contains sample information

formula

A linear model that should be used for the design matrix

min_cells

A numeric value indicating the minimum number of cells in a cluster that should be included in the analysis

Value

A dataframe containing the results of the t-test

Examples

if (FALSE) pnp_ctrl_csf_sex_age <- propellerCalc(
                                      seu_obj1 = sc_final,
                                     condition1 = "CSF_PNP",
                                     condition2 = "CSF_CTRL", 
                                     cluster_col = "cluster",
                                     meta_col = "tissue_level1",
                                     lookup = propeller_lookup,
                                     sample_col = "patient",
                                     lookup_col = "tissue_level1_CSF",
                                     min_cells = 30)
 # \dontrun{}