enrichment analysis

wrapper function to perform enrichment analysis with enrichr and save results

Usage

enrichrRun(sheet, filename, dbs, fc_thresh = 1, p_thresh = 0.001, remove_rp_mt)

Arguments

sheet

sheet name in excel file

filename

name of deg file file without extension

dbs

name of the enrichr libraries

fc_thresh

log fc threshold (default 1)

p_thresh

p value threshold (default 0.001)

remove_rp_mt

remove ribosomal and mitochondrial genes? (boolean value)

Value

save enrichrment analysis in excel sheet with multiple sheets in the folder results/enrichr

Examples

if (FALSE) enrichrFun(filename = "de_ALZ_Naive_blood", dbs = dbs, fc_thresh = 1, p_thresh = 0.001, sheet = "pDC", remove_rp_mt = TRUE) # \dontrun{}